+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`-` Header images are now used as fallback for social media preview images if there's no logo.
 - :bug:`-` Events with per-submission-type questions sometimes saw empty questions pages in the CfP flow.
 - :feature:`-` Organisers can now assign reviewers to proposals in bulk, by uploading a JSON file.
 - :feature:`-` Two new signals: ``agenda.html_above_session_pages`` and ``agenda.html_below_session_pages`` allow plugins to add custom text to the public schedule session pages.